Controls Remapping: imported template from other project
Each control must be configured for (1) main binding button (2) alt binding button (3) reset button.

There will also be a reset all button.

The remap operation is canceled (replaced with NO binding) if left mouse or escape key is pressed. In case of a successful new binding, checks are made for an override, unbinding any other controls that had an identical binding as this is not permitted