Player Position Loader: respawning / resuming at terminals
Terminal now contains spawn pos reference in front of the terminal facing the screen.

LoadPoint defaults to -1 and must be reset to -1 when loading scenes through 'resume' or 'respawn'. Non '-1' values indicate to use the load point transforms in the scene

When load point is -1: searches for terminal at saved terminal index to spawn player in front of (if this index is also -1 then default to first load point - i.e. for station)

now also loads the location AND rotation of load spots (and terminal transforms)