Floor2: created & configured wire box puzzles
this floor only has 2 as I am planning to have 5 total on floor 1 (2 of which are entirely optional, and 1 of which just unlocks a shortcut)

the janitor box uses the default puzzle as I thought that might be a good difficulty for that phase of the game? (for reference the orchard box is intended to be the HARDEST wire box puzzle in the game as the final mandatory box.

In total there should be I think at least 3 mandatory boxes? (both on floor2 and the tutorial security box?) and 4 optional/shortcut boxes?