Hangar: Lower Creature Zone
this zone is pretty much a 'prevent the player from doing much zone' . Creature spawns on enter of lower floor and will start pursuit if player moves much at all from the door.

Fun fact: it is possible to grab the keycard and get out using a generous amount of stuns if you know exactly where you need to go to get the keycard. I don't think this is realllly an issue - if anything its just a speedrun skip